+++ .coveragerc
@@ -1,7 +1,7 @@
 [run]
 omit=
-    pymdownx/pymdown.py
+    pymdownx/spoilers.py
 
 [report]
 omit=
-    pymdownx/pymdown.py
+    pymdownx/spoilers.py
