+++ docs/src/js/details.js
