+++ docs/src/scss/_details.scss
@@ -53,7 +53,7 @@
   }
 }
 
-/* Material styled spoilers */
+/* Material styled details */
 .md-typeset {
   details {
     @each $names, $appearance in (
