+++ docs/src/scss/extra.scss
@@ -9,4 +9,4 @@
 @import "inlinecode";
 @import "headeranchors";
 @import "progressbar";
-@import "spoilers";
+@import "details";
