+++ tests/extensions/details/details.html
@@ -1,0 +1,1 @@
+<details><summary>details</summary><p>content</p><details open="open"><summary>open nested details</summary><p>content</p><p>more content</p></details></details><details class="optional-class"><summary>details with class</summary><p>content</p></details>
