--- tests/extensions/spoilers/spoilers.html
@@ -1,1 +1,0 @@
-<details class="spoilers"><summary>spoiler</summary><p>content</p><details class="spoilers" open=""><summary>open nested spoiler</summary><p>content</p><p>more content</p></details></details><details class="spoilers optional-class"><summary>spoiler with class</summary><p>content</p></details>
