--- tests/extensions/spoilers/spoilers.txt
@@ -1,10 +1,0 @@
-??? "spoiler"
-    content
-
-    ???+ "open nested spoiler"
-        content
-
-        more content
-
-??? optional-class "spoiler with class"
-    content
